+++ bin/install-jsdeps.sh
@@ -35,8 +35,6 @@ if (empty($SOURCES['dependencies'])) {
 
 $CURL   = trim(`which curl`);
 $WGET   = trim(`which wget`);
-$UNZIP  = trim(`which unzip`);
-$UN7ZIP = trim(`which 7z`);
 
 if (($CACHEDIR = getenv("CACHEDIR")) && is_writeable($CACHEDIR)) {
     // use $CACHEDIR
@@ -218,12 +216,7 @@ function compose_destfile($package, $srcfile)
  */
 function extract_zipfile($package, $srcfile)
 {
-    global $UNZIP, $UN7ZIP, $CACHEDIR;
-    $phpzip = class_exists('ZipArchive', false);
-
-    if (!$phpzip && empty($UNZIP) && empty($UN7ZIP)) {
-        rcube::raise_error("PHP Zip extension, '7z' or 'unzip' programs required to install JavaScript dependencies.", false, true);
-    }
+    global $CACHEDIR;
 
     $destdir = INSTALL_PATH . $package['dest'];
     if (!is_dir($destdir)) {
@@ -234,75 +227,21 @@ function extract_zipfile($package, $srcfile)
         rcube::raise_error("Cannot write to destination directory: $destdir", false, true);
     }
 
-    // pick files from zip archive
-    if (!empty($package['pick'])) {
-        if ($phpzip) {
-            echo "Extracting files " . implode('; ', $package['pick']) . " into $destdir\n";
-            $zip = new ZipArchive;
-            if ($zip->open($srcfile) === TRUE) {
-                $zip->extractTo($destdir, $package['pick']);
-                $zip->close();
-            }
-            else {
-                rcube::raise_error("Failed to unpack " . implode('; ', $package['pick']));
-            }
-        }
-        else if (!empty($UN7ZIP)) {
-            foreach ($package['pick'] as $pattern) {
-                echo "Extracting files $pattern into $destdir\n";
-                exec(sprintf('%s x %s -y -o%s %s', $UN7ZIP, escapeshellarg($srcfile), $destdir, escapeshellarg($pattern)), $out, $retval);
-                if ($retval !== 0) {
-                    rcube::raise_error("Failed to unpack $pattern; " . implode('; ', $out));
-                }
-            }
-        }
-        else {
-            foreach ($package['pick'] as $pattern) {
-                echo "Extracting files $pattern into $destdir\n";
-                exec(sprintf('%s -o %s %s -d %s', $UNZIP, escapeshellarg($srcfile), escapeshellarg($pattern), $destdir), $out, $retval);
-                if ($retval !== 0) {
-                    rcube::raise_error("Failed to unpack $pattern; " . implode('; ', $out));
-                }
-            }
-        }
-    }
-    // unzip the archive and map source to dest files/directories
-    else if (!empty($package['map'])) {
+    if (!empty($package['map'])) {
         $extract = $CACHEDIR . '/' . $package['lib'] . '-extract';
         if (!is_dir($extract)) {
             mkdir($extract, 0774, true);
         }
+    }
 
-        if ($phpzip) {
-            $zip = new ZipArchive;
-            if ($zip->open($srcfile) === TRUE) {
-                if (!empty($package['flat'])) {
-                    for ($i = 0; $i < $zip->numFiles; $i++) {
-                        $filename = $zip->getNameIndex($i);
-                        copy("zip://" . $srcfile . "#".$filename, $extract . '/' . pathinfo($filename, PATHINFO_BASENAME));
-                    }
-                }
-                else {
-                    $zip->extractTo($extract);
-                }
-                $zip->close();
-            }
-            else {
-                rcube::raise_error("Failed to unpack $srcfile");
-            }
-        }
-        else if (!empty($UN7ZIP)) {
-            $zip_command = '%s ' . (!empty($package['flat']) ? 'e' : 'x') . ' %s -y -o%s';
-            exec(sprintf($zip_command, $UN7ZIP, escapeshellarg($srcfile), $extract), $out, $retval);
+    // pick files from the zip archive
+    rcmail_install::unzip($srcfile, $extract ?? $destdir, $package['pick'] ?? [], !empty($package['flat']));
 
-            $extract_tree = glob("$extract/*", GLOB_ONLYDIR);
-            foreach ($extract_tree as $dir) {
-                rmdir($dir);
-            }
-        }
-        else {
-            $zip_command = '%s -' . (!empty($package['flat']) ? 'j' : 'o') . ' %s -d %s';
-            exec(sprintf($zip_command, $UNZIP, escapeshellarg($srcfile), $extract), $out, $retval);
+    // map source to dest files/directories
+    if (!empty($package['map'])) {
+        $extract_tree = glob("$extract/*", GLOB_ONLYDIR);
+        foreach ($extract_tree as $dir) {
+            rmdir($dir);
         }
 
         // get the root folder of the extracted package
@@ -346,32 +285,6 @@ function extract_zipfile($package, $srcfile)
 
         // remove temp extraction dir
         exec('rm -rf ' . $extract);
-    }
-    // extract the archive into the destination directory
-    else {
-        echo "Extracting zip archive into $destdir\n";
-        if ($phpzip) {
-            $zip = new ZipArchive;
-            if ($zip->open($srcfile) === TRUE) {
-                $zip->extractTo($destdir);
-                $zip->close();
-            }
-            else {
-                rcube::raise_error("Failed to unpack $srcfile");
-            }
-        }
-        else if (!empty($UN7ZIP)) {
-            exec(sprintf('%s x %s -o%s', $UN7ZIP, escapeshellarg($srcfile), $destdir), $out, $retval);
-            if ($retval !== 0) {
-                rcube::raise_error("Failed to unzip $srcfile; " . implode('; ', $out));
-            }
-        }
-        else {
-            exec(sprintf('%s -o %s -d %s', $UNZIP, escapeshellarg($srcfile), $destdir), $out, $retval);
-            if ($retval !== 0) {
-                rcube::raise_error("Failed to unzip $srcfile; " . implode('; ', $out));
-            }
-        }
     }
 
     // remove some files from the destination